requirements.txt
openai==1.12.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
google-search-results==2.4.2
flask==3.0.2
gunicorn==21.2.0
python-json-logger==2.0.7
httpx==0.28.1 